requirements.txt
fastapi
uvicorn
pillow
ultralytics
numpy
deepface
python-multipart
tf-keras